Use newline character for concatenation in bash script
In this commit, we have revised the concatenation process for 'removable_devices_1_2' and 'other_unmounted_devices' variables in our bash script. Instead of a standard space, a newline character is now used for concatenation.

Previously, using a space for concatenation resulted in inappropriate grouping during iterations. For example, 'sdb1' and 'sda' would be together, appearing as `sdb1 sda`. To address this issue, the use of the newline character ensures each element is treated individually, like `sdb1`, `sda`.